Moved some of the MERS resolvers to more appropriate files
